LayoutTests/fast/js/function-redefinition.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="resources/js-test-pre.js"></script>
</head>
<body>
<script>
test1 = "FAIL";
var test2 = "FAIL"
function test3() {
     return "FAIL";
}
</script> 
<script>
function test1() { 
    return "SUCCESS";
}

function test2() { 
    return "SUCCESS";
}

function test3() { 
    return "SUCCESS";
}

shouldBeTrue("typeof test1 == 'function'");
shouldBeTrue("typeof test2 == 'function'");
shouldBeTrue("typeof test3 == 'function'");
shouldBe("test3()", '"SUCCESS"');
</script>
<script src="resources/js-test-post.js"></script>
</body>
</html>
