third_party/blink/web_tests/platform/generic/fast/js/function-redefinition-expected.txt

PASS typeof test1 == 'function' is true
PASS typeof test2 == 'function' is true
PASS typeof test3 == 'function' is true
PASS test3() is "SUCCESS"
PASS successfullyParsed is true

TEST COMPLETE
